Update .cabal file and src layout to eliminate linker errors.
I was getting linker errors building hsntp with Stack, I think because
the `c-sources: time.c` part was missing for the library
target. However, I was also getting warnings which seemed to be
related to everything being rebuilt for the executable targets, so I
moved the executable targets into their own source dirs and add a
library dependency.